# gea/api/openapi.base.yaml
...
openapi: 3.1.0
info:
  title: PL Re GEA API - projects, run configuration and data contracts
  version: 0.5.0
  description: "A user creates a Project, then a Run inside it. The Run is a server-side\ndraft: it is created with\
    \ a name and a treaty and filled in as the user goes\nthrough the wizard; every save is a PATCH of the fields\
    \ that changed.\nSubmitting freezes the configuration into an immutable, versioned Data\nContract (project id,\
    \ run id, ordered steps) that is stored in PostgreSQL and\ndelivered unchanged to Snowflake. Snowflake executes\
    \ the steps; this API never\ncomputes anything itself. It decides when a run may start (on its own, or in\n\
    the order and with the parallelism of its Job), shows where the execution is,\nand takes the decisions about\
    \ it: cancel it, resolve a failure.\n\nThe routes are named after what the business calls things: projects,\
    \ runs,\njobs, treaties, datasets, run parameters, submit, resolution. The fields are\nthose of workbook sheet\
    \ \"POC Data\" (`createProject.*`, `createRun.*`) under\ntheir workbook names.\n\nConventions (the same for\
    \ every operation):\n- **One envelope.** Every response body is `{\"data\": ..., \"error\": ...}` and\n  exactly\
    \ one of the two is null. On success (2xx) `data` is the resource; a\n  list is named after what it lists (`{projects,\
    \ nextCursor}`, `{runs,\n  nextCursor}`, `{regions}`). On failure (4xx, 5xx) `error` has a stable\n  `code`,\
...
servers:
...
security:
- bearerAuth: []
...
paths:
...
  /users/me:
    get:
      tags:
      - Users
      operationId: getMe
      summary: Read the caller
      description: 'The user is created the first time the identity provider''s subject is seen: in the region and
        with

        the role the provider names, or in `global` and as a Viewer. `permissions` says what the role allows.

        '
      responses:
        '200':
          description: The caller.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '304':
          $ref: '#/components/responses/NotModified'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
      parameters:
      - $ref: '#/components/parameters/IfNoneMatch'
    patch:
      tags:
      - Users
      operationId: updateMe
      summary: Change the caller's home region
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: The caller.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        default:
          $ref: '#/components/responses/Error'
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      description: 'Every user may change their own home region. The role is changed by an Admin (`PATCH /users/{userId}`).

        '
  /users:
    get:
      tags:
      - Users
      operationId: listUsers
      summary: List the users
      description: 'The people a project can be given to (`Project.Owner`), newest first, with their role and home
        region.

        Every signed-in user may read the list; it carries no e-mail address.

        '
      parameters:
      - name: role
        in: query
        description: Only the users with this role.
        schema:
          $ref: '#/components/schemas/RoleId'
      - name: region
        in: query
        description: Only the users with this home region.
        schema:
          $ref: '#/components/schemas/Region'
      - name: q
        in: query
        description: Case-insensitive match on the name.
        schema:
          type: string
          maxLength: 200
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Users, newest first.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        default:
          $ref: '#/components/responses/Error'
  /users/{userId}:
    parameters:
    - $ref: '#/components/parameters/UserId'
    get:
      tags:
      - Users
      operationId: getUser
      summary: Read a user
      x-permission: administer
      parameters:
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: The user.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '304':
          $ref: '#/components/responses/NotModified'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags:
      - Users
      operationId: updateUser
      summary: Give a user a role or another home region
      x-permission: administer
      description: 'JSON Merge Patch with `role`, `region` or both. The change is recorded in the audit log with
        who made it.

        Nobody changes their own role: that is a 403 `forbidden`, so an Admin cannot lock themselves out by mistake.

        '
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/AdministerUserRequest'
      responses:
        '200':
          description: The user.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        default:
          $ref: '#/components/responses/Error'
  /roles:
    get:
      tags:
      - Reference data
      operationId: getRoles
      summary: List the roles
      description: The roles a user can have and what each allows, in display order.
      responses:
        '200':
          description: List the roles.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
  /regions:
    get:
      tags:
      - Reference data
      operationId: getRegions
      summary: List the regions
      description: 'Project.Region: the values of createProject.region, in display order.'
      responses:
        '200':
          description: List the regions.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
  /business-purposes:
    get:
      tags:
      - Reference data
      operationId: getBusinessPurposes
      summary: List the business purposes
      description: 'Project.BusinessPurpose: the values of createProject.businessPurpose.'
      responses:
        '200':
          description: List the business purposes.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessPurposeList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
  /benefits:
    get:
      tags:
      - Reference data
      operationId: getBenefits
      summary: List the benefits
      description: 'Project.Benefit: the values of createProject.benefit.'
      responses:
        '200':
          description: List the benefits.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenefitList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
  /treaties:
    get:
      tags:
      - Reference data
      operationId: getTreaties
      summary: List the treaties a run can analyse
      description: The values of createRun.treaty. Treaties are loaded from the source system (Snowflake Module
        1); until that feed exists the list holds only what was entered by hand, and a run accepts any treaty text.
      parameters:
      - name: region
        in: query
        description: Only treaties of this region, and those that belong to no region.
        schema:
          $ref: '#/components/schemas/Region'
      - name: q
        in: query
        description: Case-insensitive match on the treaty id or name.
        schema:
          type: string
          maxLength: 200
      responses:
        '200':
          description: List the treaties a run can analyse.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreatyList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        default:
          $ref: '#/components/responses/Error'
  /datasets:
    get:
      tags:
      - Reference data
      operationId: getDatasets
      summary: List the datasets a run can use
      description: The values of createRun.dataScope.
      responses:
        '200':
          description: Datasets.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
  /projects:
    get:
      tags:
      - Projects
      operationId: listProjects
      summary: List projects
      parameters:
      - name: businessPurpose
        in: query
        schema:
          $ref: '#/components/schemas/BusinessPurpose'
      - name: region
        in: query
        schema:
          $ref: '#/components/schemas/Region'
      - name: state
        in: query
        schema:
          $ref: '#/components/schemas/ProjectState'
      - name: cycle
        in: query
        description: Project cycle (Project.Period), exact match.
        schema:
          type: string
          maxLength: 50
      - name: q
        in: query
        description: Case-insensitive match on the project name.
        schema:
          type: string
          maxLength: 200
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Projects, newest first.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags:
      - Projects
      operationId: createProject
      summary: Create a project
      x-permission: prepare
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created, or the stored answer of a repeated Idempotency-Key.
          headers:
            Location:
              $ref: '#/components/headers/Location'
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        default:
          $ref: '#/components/responses/Error'
  /projects/{projectId}:
    parameters:
    - $ref: '#/components/parameters/ProjectId'
    get:
      tags:
      - Projects
      operationId: getProject
      summary: Read a project
      parameters:
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Project with run counters.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '304':
          $ref: '#/components/responses/NotModified'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags:
      - Projects
      operationId: updateProject
      summary: Change a project
      x-permission: prepare
      description: 'JSON Merge Patch (RFC 7396): send only the members that change. Region and

        business purpose are fixed at creation because they select the run options.

        '
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Updated project.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        default:
          $ref: '#/components/responses/Error'
  /projects/{projectId}/run-parameters:
    parameters:
    - $ref: '#/components/parameters/ProjectId'
    get:
      tags:
      - Projects
      operationId: getRunParameters
      summary: List the run fields with the values, defaults and rules that apply in this project
      description: 'Every createRun field of workbook sheet "POC Data", in workbook order, with its display name,
        step, type, control and required flag. A select or multiselect field also has `options`. The options can
        depend on the project''s region, business purpose and benefits, and on the investigation: pass `investigation`
        to narrow them. When exactly one value applies it is returned as `default` (a pre-populated value). This
        is the methodology profile of the workbook, for one project.'
      parameters:
      - name: investigation
        in: query
        description: Value of createRun.investigation, for options that depend on it.
        schema:
          type: string
          maxLength: 200
      responses:
        '200':
          description: The run fields and their options.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunParameters'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /runs:
    get:
      tags:
      - Runs
      operationId: listRuns
      summary: List runs
      parameters:
      - name: projectId
        in: query
        description: Only the runs of this project.
        schema:
          type: string
          format: uuid
      - name: jobId
        in: query
        description: Only the runs of this job; `jobOrdinal` is their order of execution.
        schema:
          type: string
          format: uuid
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/RunStatus'
      - name: q
        in: query
        description: Case-insensitive match on run name, treaty or region, or the start of the run id.
        schema:
          type: string
          maxLength: 200
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Runs, newest first.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags:
      - Runs
      operationId: createRun
      summary: Create a run (the draft)
      x-permission: prepare
      description: createRun. `projectId`, `name` and `treaty` start the draft; any other createRun prop may be
        sent with it or saved later with PATCH. With `cloneSourceId` the configuration of that run is copied first;
        props sent with the request win.
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCreate'
      responses:
        '201':
          description: Run created, or the stored answer of a repeated Idempotency-Key.
          headers:
            Location:
              $ref: '#/components/headers/Location'
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        default:
          $ref: '#/components/responses/Error'
...
  /runs/{runId}/review:
    parameters:
    - $ref: '#/components/parameters/RunId'
    get:
      tags:
      - Runs
      operationId: reviewRun
      summary: Review step - readiness, missing fields and the contract that would be submitted
      description: 'Readiness, every missing prop with its step, and a preview of the contract. The ETag is the
        run''s: send it in If-Match when submitting, so that the contract is exactly what was reviewed.'
      responses:
        '200':
          description: '`preview` has no contractId yet.'
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunReview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  headers:
    ETag:
      description: Opaque validator of this representation. Send it back in If-Match or If-None-Match.
      schema:
        type: string
...
    XRequestId:
      description: Identifier of this request, for support and log correlation.
      schema:
        type: string
  responses:
    NotModified:
      description: Unchanged since the ETag in If-None-Match; no body.
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
...